Add_Function : change prototype

+++ Add_Function.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 
-float AddFunction (float x,float y);
+double AddFunction (float x,float y);
 
 
 double AddFunction (float x,float y)
